Creates manifest.json
The Manifest JSON file makes information about the app available to the browser. It is used to provide information about the app (such as name, author, icon, and description) and to specify which files are required to run the app.

You can see this in the browser inspector under Application > Manifest.